Wrap each check in txn to avoid bad state
